SIH/cloudinary/requirements.txt
blinker==1.6.2
certifi==2023.7.22
click==8.1.7
cloudinary==1.35.0
Flask==2.3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
python-dotenv==1.0.0
six==1.16.0
urllib3==1.26.16
Werkzeug==2.3.7
